requirements.txt
google-generativeai==0.8.3
requests==2.32.3
transformers==4.47.0